translate responsive

+++ src/app/component/cart/cart.component.ts
@@ -4,7 +4,8 @@ import { ICart } from '../../core/interfaces/icart';
 import { CurrencyPipe, NgClass } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { FavoriteService } from '../../core/services/favorite.service';
-
+import swal from 'sweetalert';
+import { ToastrService } from 'ngx-toastr';
 @Component({
   selector: 'app-cart',
   standalone: true,
@@ -13,11 +14,15 @@ import { FavoriteService } from '../../core/services/favorite.service';
   styleUrl: './cart.component.scss'
 })
 export class CartComponent implements OnInit{
+
+
   cartDetails : ICart  = {} as ICart;
   numberOfCartItems:number = 0;
   cartId : string | null = null;
   clearing : boolean = false;
+
   private readonly _CartService = inject(CartService);
+  private readonly _ToastrService = inject(ToastrService);
   private readonly _FavoriteService = inject(FavoriteService);
 
 
@@ -36,20 +41,31 @@ export class CartComponent implements OnInit{
 
 
   removeProduct(id : string) : void{
+
+    swal({
+      title: "Are you sure?",
+      icon: "warning",
+      buttons: ["Cancel", "Delete"], // Use an array for button labels
+      dangerMode: true,
+    }).then((willDelete) => {
       this._CartService.removeProductFromCart(id).subscribe({
         next: (res)=>{
+          this._ToastrService.success(res.message)
           this.numberOfCartItems = res.numOfCartItems;
           this.cartId = res.cartId;
 
           this.cartDetails = res.data;
         }
       })
+    });
+
   }
 
   updateQuantity(id:string , countNumber : number){
     this._CartService.updateProductQuantity(id,countNumber ).subscribe({
       next: (res)=>{
         this.numberOfCartItems = res.numOfCartItems;
+        this._ToastrService.success("count updated")
         this.cartId = res.cartId;
         this.cartDetails = res.data;
       },
@@ -60,25 +76,36 @@ export class CartComponent implements OnInit{
   }
 
   clearCart():void {
-    this.clearing = true;
-     this._CartService.clearCart().subscribe(
-      {
-        next:(res)=>{
-          this.numberOfCartItems = 0;
+    swal({
+      title: "Are you sure?",
+      icon: "warning",
+      buttons: ["Cancel", "Delete"], // Use an array for button labels
+      dangerMode: true,
+    }).then((willDelete) => {
+      this.clearing = true;
 
-          this.clearing = false;
-          if(res.message == "success"){
-            this.cartDetails = {} as ICart;
+      this._CartService.clearCart().subscribe(
+        {
+          next:(res)=>{
+            this.numberOfCartItems = 0;
+            this._ToastrService.success(res.message)
+
+            this.clearing = false;
+            if(res.message == "success"){
+              this.cartDetails = {} as ICart;
+            }
+            console.log(res)
           }
-          console.log(res)
         }
-      }
-     )
+       )
+    })
+
   }
 
   addToFavorite(id : string) : void{
       this._FavoriteService.addToWishList(id).subscribe({
         next:(res)=>{
+          this._ToastrService.success(res.message)
 
         }
       })
