+++ src/app/component/product/product.component.ts
@@ -5,11 +5,13 @@ import { CurrencyPipe } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { FavoriteService } from '../../core/services/favorite.service';
 import { CartService } from '../../core/services/cart.service';
+import { ToastrService } from 'ngx-toastr';
+import swal from 'sweetalert';
 
 @Component({
   selector: 'app-product',
   standalone: true,
-  imports: [SlicenamePipe , CurrencyPipe , RouterLink],
+  imports: [SlicenamePipe , CurrencyPipe , RouterLink ],
   templateUrl: './product.component.html',
   styleUrl: './product.component.scss',
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -18,6 +20,7 @@ export class ProductComponent {
 
   private readonly  _cartService = inject(CartService);
   private readonly  _FavoriteService = inject(FavoriteService);
+  private readonly  _ToastrService = inject(ToastrService);
 
   @Input() isFromFavorite ?:boolean = false;
 
@@ -42,6 +45,9 @@ export class ProductComponent {
       this._cartService.addProductToCart(id).subscribe({
         next:(res)=>{
           console.log(res);
+          this._ToastrService.success(
+            res.message
+          )
         }
       })
   }
@@ -49,14 +55,29 @@ export class ProductComponent {
       this._FavoriteService.addToWishList(id).subscribe({
         next:(res)=>{
           console.log(res);
+          this._ToastrService.success(
+            res.message
+          )
         }
       })
   }
-  removeFromFavorite(id:string):void{
-    this._FavoriteService.deleteFromWishList(id).subscribe({
-      next:(res)=>{
+  removeFromFavorite(id: string): void {
+    swal({
+      title: "Are you sure?",
+      icon: "warning",
+      buttons: ["Cancel", "Delete"], // Use an array for button labels
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        this._FavoriteService.deleteFromWishList(id).subscribe({
+          next: (res) => {
+            this._ToastrService.success(res.message);
+          }
+        });
       }
-    })
+    });
   }
+
+
 
 }
